backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
alembic==1.13.1
pydantic==2.5.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
cors==1.0.1
fastapi-pagination==0.12.14
clerk-backend-api==3.0.5
httpx==0.25.2
PyJWT==2.8.0 